+++ app/src/main/java/com/codepath/apps/simpletwitterclient/networking/TwitterApplication.java
@@ -1,4 +1,4 @@
-package com.codepath.apps.simpletwitterclient;
+package com.codepath.apps.simpletwitterclient.networking;
 
 import android.content.Context;
