+++ app/src/main/res/layout/activity_timeline.xml
@@ -1,7 +1,7 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.codepath.apps.simpletwitterclient.TimelineActivity">
+    tools:context="com.codepath.apps.simpletwitterclient.activities.TimelineActivity">
 
     <ListView
         android:layout_width="wrap_content"
